services/chating-service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
psycopg2-binary
sqlalchemy
firebase-admin
httpx